Refactor single run code into Kernel Factory
To simplify the DrupalKernel and hopefully make it more confortable this starts
seperating out the global environment manipulation from the kernel boot
process.

There is still a lot of global state in Drupal and there is a lot of non Kernel
driven code that needs the Kernel and all it builds to setup this global state
so there is a new kernel method called preHandle that does all the work of
handling a request without out actually routing and handling the request.

Calling this in simpletest, authorize.php, install.php and update.php allows
us to have the entire booted kernel with all the globally available backwards
compatibility and then we can run the old code.