Delay container init and dumping
This is the sort of thing I don't like to do. This is only here for testing.
When booting a kernel for testing we need to mangle settings before the
container is fetched to we can affect the runtime. If this all happens in boot
we can't do that though.

This also fixes the Kernel tests so this works.

In a normal request this will not affect the runtime in any meaningful way.