We need a request to redirect to installer
The install redirect needs to redirect to any folder the site might be in. We
either need to boot request globals or use the request object. Since we moved
the legacy globals later, using the request is pretty logical but that
requires us to work before boot(). The good news is we can avoid using
install_goto() since it does require the request globals.