Fix rebuild logic in tests.
Rebuilding doesn't really help after install. Install actually already forced
a container rebuild. What we want is to refresh the container properties on
the test object.

Also, rebuild on the kernel wasn't rebuilding with the latest module list
which in weird cases in tests caused problems because it didn't align with
what was in the module list. specifically when the container was updated
through posting a form.